flags: add API to change realname and use it instead of set_name
flags: remove r_flag_set_name API because not consistent

the right way to change the name of a flag item is by using
r_flag_rename. the old api has been then marked as static and refactored.

flags: fix a bug in r_flag_rename because it considered the new name